Add Continuous Metric Analysis tasks for CV24x7 (#3619)
* fix PR checks

* fix url and minutes for predictive

* code format fix

+++ 79-verification/src/main/java/io/harness/app/VerificationServiceApplication.java
@@ -71,7 +71,7 @@ import javax.ws.rs.Path;
  */
 public class VerificationServiceApplication extends Application<VerificationServiceConfiguration> {
   // pool interval at which the job will schedule. But here in verificationJob it will schedule at POLL_INTERVAL / 2
-  public static final int CRON_POLL_INTERVAL = (int) TimeUnit.MINUTES.toSeconds(10); // in seconds
+  public static final long CRON_POLL_INTERVAL = TimeUnit.MINUTES.toSeconds(10); // in seconds
   private static final Logger logger = LoggerFactory.getLogger(VerificationServiceApplication.class);
   private static String APPLICATION_NAME = "Verification Service Application";
   private final MetricRegistry metricRegistry = new MetricRegistry();
